@@ -7,51 +7,50 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.17.3 <3.0.0"
+  sdk: ">=2.18.0 <3.0.0"
 
 dependencies:
-  animated_custom_dropdown: ^1.2.1
-  animations: ^2.0.2
-  another_flushbar: ^1.10.29
-  audioplayers: ^0.20.1
-  cached_network_image: ^3.2.1
-  carousel_slider: ^4.1.1
-  clock: ^1.1.0
-  cloud_firestore: ^3.1.18
-  cupertino_icons: ^1.0.4
-  firebase_auth: ^3.3.20
-  firebase_core: ^1.17.0
-  firebase_storage: ^10.2.16
+  animated_custom_dropdown: ^1.4.0
+  animations: ^2.0.7
+  another_flushbar: ^1.12.29
+  cached_network_image: ^3.2.3
+  clock: ^1.1.1
+  cloud_firestore: ^4.4.0
+  collection: ^1.17.0
+  cupertino_icons: ^1.0.5
+  firebase_auth: ^4.2.6
+  firebase_core: ^2.5.0
+  firebase_storage: ^11.0.11
   flutter:
     sdk: flutter
-  flutter_hooks: ^0.18.4
-  flutter_secure_storage: ^5.0.2
+  flutter_hooks: ^0.18.5+1
+  flutter_secure_storage: ^8.0.0
   flutter_spinkit: ^5.1.0
-  flutter_svg: ^1.1.0
-  freezed_annotation: ^2.0.3
-  golden_toolkit: ^0.13.0
-  google_fonts: ^2.3.3
-  google_sign_in: ^5.3.3
-  hooks_riverpod: ^2.0.0-dev.7
-  image_picker: ^0.8.5+3
+  flutter_svg: ^2.0.1
+  freezed_annotation: ^2.2.0
+  google_fonts: ^4.0.3
+  google_sign_in: ^5.4.4
+  hooks_riverpod: ^2.2.0
+  image_picker: ^0.8.6+1
   intl: ^0.17.0
-  json_annotation: ^4.5.0
-  lottie: ^1.3.0
-  mockito: ^5.2.0
-  path: ^1.8.1
-  path_provider: ^2.0.10
-  sentry_flutter: ^6.5.1
-  shared_preferences: ^2.0.15
-  time: ^2.1.0
+  json_annotation: ^4.8.0
+  lottie: ^2.2.0
+  path: ^1.8.2
+  path_provider: ^2.0.12
+  riverpod_annotation: ^1.2.1
+  sentry_flutter: ^6.20.1
+  shared_preferences: ^2.0.17
+  time: ^2.1.3
 
 dev_dependencies:
-  build_runner: ^2.1.11
+  build_runner: ^2.3.3
   flutter_lints: ^2.0.1
   flutter_test:
     sdk: flutter
-  freezed: ^2.0.3+1
-  json_serializable: ^6.2.0
-  very_good_analysis: ^3.0.0
+  freezed: ^2.3.2
+  json_serializable: ^6.6.1
+  riverpod_generator: ^1.2.0
+  very_good_analysis: ^4.0.0+1
 
 flutter:
   uses-material-design: true
