@@ -8,28 +8,19 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 import '../controllers/auth_controller.dart';
 
 // States
-import '../../shared/states/future_state.codegen.dart';
+import '../../../global/states/states.dart';
 
 // Helpers
-import '../../../helpers/constants/app_assets.dart';
-import '../../../helpers/constants/app_colors.dart';
-import '../../../helpers/constants/app_styles.dart';
-import '../../../helpers/constants/app_typography.dart';
-
-// Routing
-import '../../../config/routes/app_router.dart';
+import '../../../helpers/constants/constants.dart';
 
 // Widgets
-import '../../shared/widgets/custom_circular_loader.dart';
-import '../../shared/widgets/custom_dialog.dart';
-import '../../shared/widgets/custom_text_button.dart';
+import '../../../global/widgets/widgets.dart';
 
 class LoginScreen extends HookConsumerWidget {
   const LoginScreen({super.key});
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-
     ref.listen<FutureState<bool?>>(
       authControllerProvider,
       (_, authState) => authState.whenOrNull(
@@ -58,20 +49,17 @@ class LoginScreen extends HookConsumerWidget {
             Insets.gapH10,
 
             // Logo
-            Text(
+            CustomText.heading(
               'Family Expense Tracker',
-              style: AppTypography.primary.heading34.copyWith(
-                fontSize: 45,
-              ),
+              fontSize: 45,
             ),
 
             Insets.gapH20,
 
             Align(
               alignment: Alignment.centerLeft,
-              child: Text(
+              child: CustomText.title(
                 'Save your money with\nconscious spending',
-                style: AppTypography.primary.title18,
               ),
             ),
 
@@ -80,7 +68,7 @@ class LoginScreen extends HookConsumerWidget {
             // Login Button
             CustomTextButton.gradient(
               width: double.infinity,
-              gradient: AppColors.buttonGradientPurple,
+              gradient: AppColors.buttonGradientPrimary,
               onPressed: () {
                 ref.read(authControllerProvider.notifier).loginWithGoogle();
               },
@@ -94,12 +82,10 @@ class LoginScreen extends HookConsumerWidget {
                     orElse: () => child!,
                   );
                 },
-                child: Center(
-                  child: Text(
+                child: const Center(
+                  child: CustomText(
                     'GOOGLE SIGN IN',
-                    style: AppTypography.secondary.body16.copyWith(
-                      color: Colors.white,
-                    ),
+                    color: Colors.white,
                   ),
                 ),
               ),
